refactor: controller, service 대신 Application -> BowlingGame으로 수정
- View 와 비즈니스 로직을 결국에 한 곳에서 관리하게 되어 Application에서 관리
